requirements.txt
openai
anthropic
google-genai
datasets
torch
transformers
pandas
numpy
matplotlib
tqdm
nltk
scikit-learn
sentence-transformers
python-dotenv
flask
pymongo
pyyaml
tiktoken
scipy
statsmodels 
vllm 